working through record component conversion
